语言翻译和切换

+++ app/src/main/java/com/smona/btwriter/SplashActivity.java
@@ -43,7 +43,6 @@ public class SplashActivity extends BaseLanguageActivity {
             ARouterManager.getInstance().gotoActivity(path);
             overridePendingTransition(0, 0);
             finish();
-        }, (long) (1000));
+        }, 1000);
     }
-
 }
